% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_list.r
\name{as.typed_list}
\alias{as.typed_list}
\title{Typed list}
\usage{
as.typed_list(x, type_class)
}
\arguments{
\item{x}{input object}

\item{type_class}{generic class name of elements in the list}
}
\value{
typed list with appropriate type assigned
}
\description{
Typed list ensures all items in the list are of the same type.
However, ducktyping takes place, which is if the element has target class
it doesn't mean that it has the same structure as other items in the list.
}
