% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{c_extended}
\alias{c_extended}
\title{Extended vector creation function}
\usage{
c_extended(..., as.names = TRUE, as.list = FALSE)
}
\arguments{
\item{...}{items to vectorize}

\item{as.names}{make names from items in a vector}

\item{as.list}{return resulting object as a list}
}
\description{
Extended vector creation function
}
