% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_list.r
\name{$.typed_list}
\alias{$.typed_list}
\title{S3 method for '$' function for typed_list}
\usage{
\method{$}{typed_list}(x, name)
}
\arguments{
\item{x}{object under test}

\item{name}{property name}
}
\description{
S3 method for '$' function for typed_list
}
