% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.r
\name{defaultIf}
\alias{defaultIf}
\title{Default-if function}
\usage{
defaultIf(.predicate, .value, .default)
}
\arguments{
\item{.predicate}{Function to use for evaluation for returning default}

\item{.value}{value to evaluate with predicate}

\item{.default}{default value to return if .value has not passed predicate check}
}
\description{
Default-if function
}
