% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.r
\name{\%$$\%}
\alias{\%$$\%}
\title{Extract parts of the typed_list object and convert them to typed_list for further extraction}
\usage{
lhs \%$$\% rhs
}
\arguments{
\item{lhs}{Input object}

\item{rhs}{Part of the object}
}
\value{
list of parts of input colleciton of objects
}
\description{
Extract parts of the typed_list object and convert them to typed_list for further extraction
}
