% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.r
\name{\%??\%}
\alias{\%??\%}
\title{Double-question S3 generic operator}
\usage{
lhs \%??\% rhs
}
\arguments{
\item{lhs}{left argument}

\item{rhs}{right argument}
}
\description{
Double-question S3 generic operator
}
