% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_list.r
\name{is_typed_list}
\alias{is_typed_list}
\title{Typed list class check}
\usage{
is_typed_list(x)
}
\arguments{
\item{x}{object under test}
}
\value{
TRUE or FALSE
}
\description{
Check whether object is of class 'typed_class'
}
