% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.r
\name{throw_exception}
\alias{throw_exception}
\title{Wrapper for stop function}
\usage{
throw_exception(arg, message = NULL)
}
\arguments{
\item{arg}{argument which must be a part of exception}

\item{message}{additional message to exception}
}
\description{
Wrapper for stop function
}
