% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_list.r
\name{typed_list.is}
\alias{typed_list.is}
\title{Check typed_list generic type of items}
\usage{
typed_list.is(x, type_class)
}
\arguments{
\item{x}{object under test}

\item{type_class}{class name to check for generic type}
}
\value{
TRUE or FALSE
}
\description{
Check typed_list generic type of items
}
