% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_list.r
\name{typed_list.typeof}
\alias{typed_list.typeof}
\title{Get generic type of items in the 'typed_list'}
\usage{
typed_list.typeof(x)
}
\arguments{
\item{x}{object under test}
}
\value{
generic type of items
}
\description{
Get generic type of items in the 'typed_list'
}
