% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{validate_classes}
\alias{validate_classes}
\title{Validate that object contains at least one class from provided list of classes}
\usage{
validate_classes(object, ..., classes = character())
}
\arguments{
\item{object}{object under test}

\item{...}{dots argument for classes if empty}

\item{classes}{vector of classes}
}
\value{
input object
}
\description{
Validate that object contains at least one class from provided list of classes
}
