% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_list.r
\name{validate_typed_list_class}
\alias{validate_typed_list_class}
\title{Validate that object has typed list class}
\usage{
validate_typed_list_class(x)
}
\arguments{
\item{x}{object under test}
}
\description{
Validate that object has typed list class
}
