% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{with_class}
\alias{with_class}
\title{Extend class of object}
\usage{
with_class(obj, className)
}
\arguments{
\item{obj}{object under test}

\item{className}{class to append to list of classes}
}
\description{
Extend class of object
}
